Fix compilation on Windows 64 bits
Code usually uses std functions which returns a size_t and size_t
is not going to fit into an int on 64 bits. Today it triggers a
warning of potential data loss.

We could disable the warning but it's better to fix the code.
Please note that we have various casts, these are mandatory (though
unfortunate) at the moment because the API in Chromium requires
explicitely an int for example. This can be fixed upstream and
bugs have been reported (e.g. crbug.com/167187).

+++ application/common/manifest_handler.cc
@@ -211,7 +211,7 @@ void ManifestHandlerRegistry::ReorderHandlersGivenDependencies() {
       ManifestHandler* handler = *iter;
       const std::vector<std::string>& prerequisites =
           handler->PrerequisiteKeys();
-      int unsatisfied = prerequisites.size();
+      size_t unsatisfied = prerequisites.size();
       for (size_t i = 0; i < prerequisites.size(); ++i) {
         ManifestHandlerMap::const_iterator prereq_iter =
             handlers_.find(prerequisites[i]);
